requirements.txt
Flask==2.0.1
botbuilder-core==4.11.0
botbuilder-schema==4.11.0
botframework-connector==4.11.0
azure-functions